requirements.txt
Flask==3.1.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.182.0
Werkzeug==3.1.3